+++ CHANGELOG.rst
@@ -5,5 +5,6 @@ Unreleased
 * feat: ``Introduction.ombud_messages``
 * feat: ``general.boards``
 * feat: ``general.cs_codes``
+* feat: ``general.industrial_groups``
 * fixed: ``database.add_instrument_to_db``
 * fixed!: ``beautifulsoup4`` and ``lxml`` are now required as dependencies.